Move the rest of the view to shared
