+++ appshared/src/main/java/io/reark/rxgithubapp/shared/view/RepositoriesView.java
@@ -23,7 +23,7 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-package io.reark.rxgithubapp.advanced.view;
+package io.reark.rxgithubapp.shared.view;
 
 import android.content.Context;
 import android.support.annotation.NonNull;
@@ -42,7 +42,7 @@ import java.util.List;
 
 import io.reark.reark.utils.Preconditions;
 import io.reark.reark.utils.RxViewBinder;
-import io.reark.rxgithubapp.R;
+import io.reark.rxgithubapp.shared.R;
 import io.reark.rxgithubapp.shared.pojo.GitHubRepository;
 import io.reark.rxgithubapp.shared.viewmodels.RepositoriesViewModel;
 import io.reark.rxgithubapp.shared.viewmodels.RepositoriesViewModel.ProgressStatus;
