+++ appshared/src/main/res/layout/repositories_fragment.xml
@@ -8,7 +8,7 @@
               android:paddingTop="@dimen/activity_vertical_margin"
               android:paddingBottom="@dimen/activity_vertical_margin">
 
-    <io.reark.rxgithubapp.advanced.view.RepositoriesView
+    <io.reark.rxgithubapp.shared.view.RepositoriesView
             android:id="@+id/repositories_view"
             android:layout_width="match_parent"
             android:layout_height="match_parent">
@@ -33,5 +33,5 @@
                 android:layout_height="match_parent"
                 android:layout_marginTop="66dp"/>
 
-    </io.reark.rxgithubapp.advanced.view.RepositoriesView>
+    </io.reark.rxgithubapp.shared.view.RepositoriesView>
 </LinearLayout>
